+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix issue where dropdowns were not in front of icons.
 - Update transaction approval styles.
 - Align failed and successful transaction history text.
+- Fix issue where large domain names and large transaction values would misalign the transaction history.
 
 ## 2.5.0 2016-06-29